+++ encoding/thrift/thriftrw-plugin-yarpc/client.go
@@ -37,16 +37,18 @@ package <$pkgname>
 <$yarpc     := import "go.uber.org/yarpc">
 <$transport := import "go.uber.org/yarpc/api/transport">
 <$thrift    := import "go.uber.org/yarpc/encoding/thrift">
-<$context   := import "context">
+
+</* Note that we import things like "context" inside loops rather than at the
+    top-level because they will end up unused if the service does not have any
+    functions.
+ */>
 
 // Interface is a client for the <.Service.Name> service.
 type Interface interface {
-	<if .Parent>
-		<$parentPath := printf "%s/%sclient" .ParentModule.ImportPath (lower .Parent.Name)>
-		<import $parentPath>.Interface
+	<if .Parent><import .ParentClientPackagePath>.Interface
 	<end>
-
 	<range .Service.Functions>
+		<$context := import "context">
 		<.Name>(
 			ctx <$context>.Context, <range .Arguments>
 			<.Name> <formatType .Type>,<end>
@@ -64,10 +66,13 @@ type Interface interface {
 //
 // 	client := <$pkgname>.New(dispatcher.ClientConfig("<lower .Service.Name>"))
 func New(c <$transport>.ClientConfig, opts ...<$thrift>.ClientOption) Interface {
-	return client{c: <$thrift>.New(<$thrift>.Config{
-		Service: "<.Service.Name>",
-		ClientConfig: c,
-	}, opts...)}
+	return client{
+		c: <$thrift>.New(<$thrift>.Config{
+			Service: "<.Service.Name>",
+			ClientConfig: c,
+		}, opts...),
+		<if .Parent> Interface: <import .ParentClientPackagePath>.New(c),
+		<end>}
 }
 
 func init() {
@@ -76,11 +81,16 @@ func init() {
 	})
 }
 
-type client struct{ c <$thrift>.Client }
+type client struct {
+	<if .Parent><import .ParentClientPackagePath>.Interface
+	<end>
+	c <$thrift>.Client
+}
 
 <$service := .Service>
 <$module := .Module>
 <range .Service.Functions>
+<$context := import "context">
 <$prefix := printf "%s.%s_%s_" (import $module.ImportPath) $service.Name .Name>
 
 func (c client) <.Name>(