+++ internal/crossdock/thrift/gauntlet/secondserviceserver/server.go
@@ -75,7 +75,10 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 			},
 		},
 	}
-	return thrift.BuildProcedures(service, opts...)
+
+	procedures := make([]transport.Procedure, 0, 2)
+	procedures = append(procedures, thrift.BuildProcedures(service, opts...)...)
+	return procedures
 }
 
 type handler struct{ impl Interface }